Visualizacion de datos

+++ database/migrations/2022_04_10_200835_create_clientes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('nombre');
             $table->string('apellido');
             $table->string('documento_tipo');
-            $table->integer('documento_numero');
+            $table->bigInteger('documento_numero');
             $table->string('correo');
             $table->string('telefono');
             $table->string('direccion');
